Milestone 2 Phase 2/deployment/requirements.txt
waitress
numpy
Pillow
streamlit
tensorflow